Unify all four dispatchers
